refactor(edge-stacks): move config inspect to endpoints handler
