+++ api/http/handler/edgestacks/handler.go
@@ -14,7 +14,6 @@ type Handler struct {
 	*mux.Router
 	requestBouncer   *security.RequestBouncer
 	EdgeStackService portainer.EdgeStackService
-	EndpointService  portainer.EndpointService
 	FileService      portainer.FileService
 	GitService       portainer.GitService
 }
@@ -39,7 +38,5 @@ func NewHandler(bouncer *security.RequestBouncer) *Handler {
 		bouncer.AdminAccess(httperror.LoggerHandler(h.edgeStackFile))).Methods(http.MethodGet)
 	h.Handle("/edge_stacks/{id}/status",
 		bouncer.AdminAccess(httperror.LoggerHandler(h.edgeStackStatusUpdate))).Methods(http.MethodPut)
-	h.Handle("/edge_stacks/{id}/config",
-		bouncer.AdminAccess(httperror.LoggerHandler(h.edgeStackInspectConfig))).Methods(http.MethodGet)
 	return h
 }
